requirements.txt
fastapi
uvicorn
playwright
openai
google-generativeai
python-dotenv